move lógica de csv para trait
